Adapt VectorBQM to use Vector
